新政中间件

+++ model/house.go
@@ -1,9 +1,9 @@
 package model
 
 type House struct {
-	ID    int     `json:"id"`
-	Name  string  `json:"name"`
-	Price float64 `json:"price"`
+	ID    int    `json:"id"`
+	Name  string `json:"name"`
+	Price int    `json:"price"`
 }
 
 func (h *House) TableName() string {
